Create Note and input
For create note I have a usestate isShowing to decide whether the state is in the first state or if a user has begun to create a note. This decides if they see just the note text or the title and close button. THe input is very simple and just taking in props className placeHolder and onClick at the moment.

Will need to create a button component.

Have been taking a look at the design for site, and I do also need a circular button. Need to decide if this will be the same component or different.